+++ src/clj/clj_toolbox/string_tools.clj
@@ -1,6 +1,0 @@
-(ns clj-toolbox.string-tools
-  (:require
-    [clojure.string :as str]
-    [clj-toolbox.strings :as strings]))
-
-(def ^:deprecated box-trim strings/box-trim)